refactor code in edit buttons
